Move AbtractIbftMessageData to common types
