+++ consensus/common/build.gradle
@@ -33,6 +33,7 @@ dependencies {
   implementation project(':crypto')
   implementation project(':ethereum:api')
   implementation project(':ethereum:core')
+  implementation project(':ethereum:p2p')
   implementation project(':ethereum:rlp')
   implementation project(':util')
