+++ consensus/ibft/src/main/java/org/hyperledger/besu/consensus/ibft/messagedata/ProposalMessageData.java
@@ -14,12 +14,13 @@
  */
 package org.hyperledger.besu.consensus.ibft.messagedata;
 
+import org.hyperledger.besu.consensus.common.bft.messages.AbstractBftMessageData;
 import org.hyperledger.besu.consensus.ibft.messagewrappers.Proposal;
 import org.hyperledger.besu.ethereum.p2p.rlpx.wire.MessageData;
 
 import org.apache.tuweni.bytes.Bytes;
 
-public class ProposalMessageData extends AbstractIbftMessageData {
+public class ProposalMessageData extends AbstractBftMessageData {
 
   private static final int MESSAGE_CODE = IbftV2.PROPOSAL;
